packages/promptvault/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promptvault-ai"
version = "1.0.0"
description = "Version control and registry for prompt assets"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{ name = "Scott Converse" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "prompttools-core-ai>=1.0,<2.0",
    "typer[all]>=0.12",
    "pyyaml>=6.0",
    "rich>=13.0",
    "packaging>=23.0",
]

[project.scripts]
promptvault = "promptvault.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "ruff",
    "mypy",
]

[tool.hatch.build.targets.wheel]
packages = ["src/promptvault"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py39"
line-length = 100
